Changes made to allow code to compile with a modern windows compiler and still be compatible with Linux/Unix/OSX. Will check in a new Visual C++ 2008 directory in a separte commit.
made frequent use of,
#ifdef _MSC_VER
  <do microsoft specific stuff>
#else
  <do Linux/Unix/OSX stuff>
#endif

used static_cast<...>() to convert integer or longs to doubles in pow() and sqrt().

Added an assert(0) to the windows section of code that should never be called as it will lead to infinite recursion.

In a write() functionk, conveted some floats to doubles, but need to verify that they read in correctly, or change read code to double.

Changed a time variable from long to time_t.

+++ src/GamIO/XWinAcqus.cc
@@ -318,10 +318,12 @@ ostream& XWinAcqus::dpa(ostream& ostr) const
   {
   ostr << "\n" << string(29, ' ') << "Acquisition Parameters";
   ostr << "\n" << string(79, '=') << "\n\n";
-  bru(ostr, "Date_",    _DATE,         "",                           0);
-bru(ostr, "Time",     "NOCH NICHT",         "",                           1);
+	// *** _DATE has been redefined as a time_t.  
+	// Cast as long on output to make sure no loss of data.
+  bru(ostr, "Date_",    static_cast<long>(_DATE),         "",        0);
+  bru(ostr, "Time",     "NOCH NICHT",         "",                    1);
 // int tf = pset.getString("Time",sval2);
-//  bru(ostr, "Time",     _TIME,         "",                           1);
+//  bru(ostr, "Time",     _TIME,         "",                         1);
   bru(ostr, "PULPROG",  _PULPROG,      "",                           0);
   bru(ostr, "AQ_mod",   AQ_modS(),     "",                           1);
   bru(ostr, "TD",       TD(),          "",                           0);
